fix(cv): align range AP test with inclusive IoU and clear CodeQL alerts
The inclusive-threshold fix (a match is IoU at or above the threshold,
as in pycocotools) also applies to the zero endpoint of an AP range, as
the review noted. Range_ZeroOverlapIsNotAMatchEvenAtZeroThreshold still
asserted the old behaviour and failed Unit 13; it now expects 1/11 over
the 0..1 step 0.1 grid, where only threshold 0 matches.

CodeQL flagged integer products widened to double after multiplying
(YOLOHead scales, CRNN CTC weights) and an intentional integer division
in the TaskAligned anchor grid. Promote before multiplying; name the
row/column split. Results are bit-identical where nothing overflowed.

+++ src/ComputerVision/OCR/Recognition/CRNN.cs
@@ -637,7 +637,7 @@ public partial class CRNN<T> : OCRBase<T>
         var weights = new Tensor<T>(new[] { labels.Length });
         for (int b = 0; b < labels.Length; b++)
         {
-            weights[b] = NumOps.FromDouble(1.0 / (Math.Max(1, labels[b].Length) * labels.Length));
+            weights[b] = NumOps.FromDouble(1.0 / ((double)Math.Max(1, labels[b].Length) * labels.Length));
         }
 
         return Engine.ReduceSum(Engine.TensorMultiply(perSequence, weights), null);
